Game Scenes
Added the scenes for the various states in the game, including the main menu scene, the level scenes, and the scenes that pop up when you win or lose in the game. Also included basic UI elements for the menu, win, and lose scenes. Finally, added some placeholder art to be used until final art can be included.