Levels Connected
Created a Script called LevelControl that loads the first level and the title screen when the correct buttons are clicked. Also modified the player controller so that the player will be able to move between levels. In addition, tags were added for each of the triggers that load the corresponding levels, as well as the win scene. Finally, added all game scenes to the build settings so that the scene manager functions can work properly.